Libraries with backspace message and expansion file!
Source b699f5f22b4bc15622502e04f1018b4bc3bb0e17

+++ ios/include/private/PatternRecognizer/FLBlackBox.h
@@ -27,7 +27,7 @@
 #define BASEPOINT_CUSTOM2_ID (BASEPOINT_CUSTOM1_ID + 1)
 
 
-typedef uint64_t BBKey;
+typedef uint32_t BBKey;
 typedef uint16_t BBValue;
 
 #if 1
